feat: 增加音视频数据规范 (#85)
* feat(mdedia): 增加音视频规范

* chore(mdeida): 优化注释

* style: 调整音视频类型规范结构

---------
