+++ src/index.ts
@@ -1,6 +1,6 @@
 export * from './type';
 export * from './composition';
-export * from './image';
+export * from './texture';
 export * from './constants';
 export * from './number-expression';
 export * from './scene';
@@ -17,6 +17,8 @@ export * from './item/sprite-item';
 export * from './item/spine-item';
 export * from './item/effect-item';
 export * from './item/text-item';
+export * from './item/video-item';
+export * from './item/audio-item';
 export * from './item/model';
 export * from './vfx-item-data';
 export * from './animation-clip-data';
@@ -26,3 +28,4 @@ export * from './components';
 export * from './vfx-item-data';
 export * from './buitin-object-guid';
 export * from './timeline';
+export * from './assets';
